third_party/blink/web_tests/http/tests/security/mixedContent/strict-mode-websocket-blocked.https.html

<!DOCTYPE html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="block-all-mixed-content">
    <script src="/js-test-resources/testharness.js"></script>
    <script src="/js-test-resources/testharnessreport.js"></script>
</head>
<body>
    <script>
        test(function () {
            assert_throws_dom("SecurityError",
                              function () {
                                  new WebSocket("ws://example.test:8880/echo");
                              });
        }, "Test that insecure websocket construction throws.");
    </script>
</body>
</html>
