third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/mixedContent/strict-mode-websocket-blocked.https-expected.txt

CONSOLE ERROR: line 13: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/strict-mode-websocket-blocked.https.html' was loaded over HTTPS, but attempted to connect to the insecure WebSocket endpoint 'ws://example.test:8880/echo'. This request has been blocked; this endpoint must be available over WSS.
This is a testharness.js-based test.
PASS Test that insecure websocket construction throws. 
Harness: the test ran to completion.
